ABCOpts: releases can have memory unsafe side effects due to the deinit method
To guarantuee memory safety we need to guarantuee that the deinit method can not
do anything unsafe. This means we need to treat general releases like any
arbitrary function call. We need to assume it does something unsafe.
Conservatively, we only releases on arrays are known to be safe.

Thanks Andy for pointing this out to me!

Swift SVN r23181