Fix check related to null conformances.
Verify that existentials have null conformances, just as is required of
archetypes, and require non-null conformances in other cases.

The last require that we had here attempted to allow for null
conformances, but we check just above that the conformance is non-null,
and this was firing on existentials with some other changes that I have.

Swift SVN r23200