Fix for rdar://problem/18926814 (Fuzzing swift: Swift parser crash).
The issue is reproducible in erroneous code scenario when a string literal
has invalid stuff inside interpolation segment, like: “... \( abc } ) ...”.

In this case we used to do some magic for switching context inside the string 
and parse the interior of \(...) as regular expression list, but expected that 
the parsing finishes at closing “)” which does not necessarily true in case
the code has errors.

The assertion was replaced with an error diagnostics.



Swift SVN r23296