Remove the locks from getGenericMetadata to accelerate Debug builds.
This commit removes the locks from the family of getXXXXMetadata APIs in the
fast path that does not create a new metadata. We still need the lock for the
cache-miss case because constructing metadata has side effects and we must
not construct two metadatas at the same time.

I am seeing a 25% - 30% boost in performance on most workloads that are built
with -Onone.

Swift SVN r23353