Fix for rdar://problem/18990358 (Fuzzing swift: Swift verifier crashes).
Having VarDecl created with IsStatic=false and then modifying it to true later seems 
fragile, e.g. in this bug this attribute was not properly set because of early-outs. 
In this patch I just moved code fixing VarDecl, but in future we should consider 
making this flag immutable and initializing it on creation.



Swift SVN r23406