This is a re-factored, more modular implementation of jump-threading for checked_cast_br [exact] based on Michael's comments.
The implementation is inside a dedicated class now. Huge methods are split into smaller ones that are much easier for reviewing and understanding.
The whole thing could be moved out of SimplifyCFG into a dedicated file eventually. But currently it is not done as it still has too many dependencies on the internals of SimplifyCFG.

And, BTW, this dominator-based implementation of checked_cast_br jump-threading gives a 5x boost on the relevant PolymorphicCalls benchmark.

Swift SVN r23536