[silgen] Refactor handling of self parameters so that the convention of the self parameter is changed after we know the number of arguments.
This cleans up the code and removes the circular dependency in between handling
the self parameter and computing the number of arguments. Previously we were
handling this by separately computing the number of parameters twice. This
ensures that we only use the real param count after we lower all of the
arguments helping to avoid any possibility of a bug arising due to the two
methods of computing the number of parameters differing.

This does not have any changes to SILGen output since it is just a refactoring.

This also makes it very easy to change the self parameter convention to
guaranteed by just changing the convention returned by the default swift
convention (the reason why I am making this change).

Swift SVN r23621