Disable OptimizationOptions-without-stdlib-checks
Because the bots build a release+asserts compiler and an unchecked stdlib, this
test is not currently applicable. The checked version of this test was disabled
but this test should've been disabled as well.

Original commit message r23074:

Fix the bots by requiring two options on this test that never can co-exist as true.

This is to work around a bug on the unchecked optimized stdlib builder which due
to a problem in the cmake build chain runs the checks enabled version of this
test. This commit makes sure that the test will not run until this is resolved.

Swift SVN r23624