Fix leak caused by allocbox-to-stack.
In allocbox-to-stack we were cloning closures to remove box parameters
but in the process we were also removing the release on the box. We
still need to release it so that if it is the last reference anything it
owns gets released.

This change inserts the release before the partial_apply.

Fixes rdar://problem/17872402.

Swift SVN r23876