Get rid of a memory allocation on each call to the getXXXMetadata functions.
We were using std::function that allocated memory. Unfortunately c++ lambdas
don't have a type so I had to make the function findOrAdd templated on the
type of the callback. This is not a big deal since the class is already
templated and there is a single call site.

Swift SVN r23919