[func-sig-opts] The projection tree knows how to explode address arguments... but the rest of func sig opts doesn't. So do not try to optimize in such cases.
I do not think that function signature opts is the right place to handle
promotion of @in, @out arguments. The right place to do such things is when we
perform specialization. That is the natural place to do it since that is where
we transition from address only to loadable types.

Swift SVN r23940